+++ examples/v2_get_block_transaction_events.rs
@@ -47,6 +47,7 @@ async fn main() -> anyhow::Result<()> {
             while let Some(event) = events.next().await.transpose()? {
                 if event
                     .affected_contracts()
+                    .known_or_err()?
                     .contains(&ContractAddress::new(866, 0))
                 {
                     println!(